requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
psutil>=5.9.0 